Move ChatPane to bootstrap-chat
